+++ sphinx/ext/autosummary/generate.py
@@ -274,7 +274,10 @@ def members_of(obj: Any, conf: Config) -> Sequence[str]:
     if conf.autosummary_ignore_module_all:
         return dir(obj)
     else:
-        return getall(obj) or dir(obj)
+        all = getall(obj)
+        # if __all__ is not set, return dir, but __all__ might also be an
+        # empty list, so do not use `return getall(obj) or dir(obj)`.
+        return dir(obj) if all is None else all
 
 
 def generate_autosummary_content(